pyproject.toml
[tool.poetry]
name = "sistema-lanchonete"
version = "1.0.0"
description = "Sistema de Gerenciamento de Lanchonete - Controle de estoque e vendas"
authors = ["Victor Soares Ferreira <victor@exemplo.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
tabulate = "^0.9.0"
openpyxl = "^3.1.0"
requests = "^2.31.0"
pandas = "^2.0.0"
matplotlib = "^3.7.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
nuitka = "^1.8.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
lanchonete = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=src --cov-report=html --cov-report=term-missing"

# Configuração para empacotamento com Nuitka
[tool.nuitka]
main = "main.py"
output-filename = "SistemaLanchonete.exe"
onefile = true
enable-plugin = "tk-inter"
windows-disable-console = true
windows-icon-from-ico = "assets/icon.ico"
include-data-dirs = ["data=data", "assets=assets"]
